Move parse-complex-response to shared.
